contact-manager/requirements.txt
Flask==3.1.2
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
numpy==2.3.3
pandas==2.2.2